Fix attempt to kill VARA modems during install if user un-checks "Start now"

+++ CORE/VARA/30_VARA.sh
@@ -58,12 +58,16 @@ install_vara () {
 unzip -o -d /tmp $MODULE_DIR/"VARA FM v4.3.8 setup.zip"
 wine /tmp/"VARA FM setup (Run as Administrator)".exe
 sleep 0.5
-killall VARAFM.exe
+if pidof VARAFM.exe; then
+	killall VARAFM.exe
+fi
 
 unzip -o -d /tmp $MODULE_DIR/"VARA HF v4.8.9 setup.zip"
 wine /tmp/"VARA setup (Run as Administrator)".exe
 sleep 0.5
-killall VARA.exe
+if pidof VARA.exe; then
+	killall VARA.exe
+fi
 }
 
 config_vara () {
